refactor: move Hit and HitWithInsights.js into their own file
