+++ React InstantSearch/algolia-insights/src/App.js
@@ -1,26 +1,21 @@
 import algoliasearch from 'algoliasearch/lite';
 import React, { Component } from 'react';
 import {
-  InstantSearch,
-  SearchBox,
-  Pagination,
-  Highlight,
   ClearRefinements,
-  RefinementList,
   Configure,
   Hits,
-  connectHitInsights,
+  InstantSearch,
+  Pagination,
+  RefinementList,
+  SearchBox,
 } from 'react-instantsearch-dom';
-
-import PropTypes from 'prop-types';
 import './App.css';
+import { HitWithInsights } from './HitWithInsights';
 
 const searchClient = algoliasearch(
   'B1G2GM9NG0',
   'aadef574be1f9252bb48d4ea09b5cfe5'
 );
-
-const HitWithInsights = connectHitInsights(window.aa)(Hit);
 
 class App extends Component {
   render() {
@@ -44,44 +39,5 @@ class App extends Component {
     );
   }
 }
-
-function Hit({ hit, insights }) {
-  return (
-    <div>
-      <img src={hit.image} align="left" alt={hit.name} />
-      <div className="hit-name">
-        <Highlight attribute="name" hit={hit} />
-      </div>
-      <div className="hit-description">
-        <Highlight attribute="description" hit={hit} />
-      </div>
-      <div className="hit-price">${hit.price}</div>
-      <button
-        className="hit-action"
-        onClick={() => {
-          insights('clickedObjectIDsAfterSearch', {
-            eventName: 'Add to favorite',
-          });
-        }}
-      >
-        Send click
-      </button>{' '}
-      <button
-        className="hit-action"
-        onClick={() => {
-          insights('convertedObjectIDsAfterSearch', {
-            eventName: 'Add to basket',
-          });
-        }}
-      >
-        Send conversion
-      </button>
-    </div>
-  );
-}
-
-Hit.propTypes = {
-  hit: PropTypes.object.isRequired,
-};
 
 export default App;
